Add todo comments for continuation
