+++ cypress/integration/specs/search-course.spec.ts
@@ -1,6 +1,5 @@
-import { resolve } from "cypress/types/bluebird";
 import { HomePage } from "../../page-objects/demo-home.page";
-
+import { CoursePage } from "../../page-objects/course.page";
 
 
 describe('Demo Course - Search for a Course', () => {
@@ -15,42 +14,10 @@ describe('Demo Course - Search for a Course', () => {
         cy.contains('Test your Selenium / QTP Scripts').should('be.visible');
     });
 
-    it.skip('List all action name', () => {
-
-        //TODO: Move to a custom command or method within the page object
-        let category:string[] = [];
-        let actions:string[] = [];
-
-        //Get all action name and store to json
-        cy.get('.linkbox').each(($category) => {
-
-            //Get category name
-            const text = $category.find('.heading').text();
-            cy.log("Category: " + text);
-            category.push(text);
-
-            //Get each action name and push to the related category
-            cy.wrap($category).find('ul li h2').each(($action) => {
-                const text = $action.text();
-                cy.log("Action: " + text);
-
-                actions.push(text);
-            }).then(() => {
-                category[text] = actions;
-                actions = [];
-            })
-
-        }).then(() =>{
-            console.log(category);
-            cy.writeFile('cypress/fixtures/actionsName.json', category);
-        }) 
-
-
-
-    });
-
 
     it.skip('Navigate to the Submit Button Clicked link', () => {
+        home.getActionName();
+
         //Arrange: Get the link
         //TODO: Create a function to fetch the link "Submit Button Clicked"
         const link = 'http://www.qa.way2automation.com/';
@@ -59,8 +26,7 @@ describe('Demo Course - Search for a Course', () => {
         cy.visit(link);
 
         //Assert
-        //TODO: Check other forms of text matchers
-        cy.url().should('include', 'way2auto_jquery');
+        cy.url().should('include.text', 'way2auto_jquery');
         home.registrationForm.find('h3')
                             .contains('Dummy Registration Form')
                             .should('be.visible');
@@ -68,7 +34,6 @@ describe('Demo Course - Search for a Course', () => {
 
 
     it.skip('Fill out the registration form', () => {
-        //TODO: Add custom command for filling up the form or a custom method within page object.
         //Arrange
         cy.fixture('registration').then(data =>{
             home.nameTxtbox.type(data.name);
@@ -83,11 +48,11 @@ describe('Demo Course - Search for a Course', () => {
         home.submitBtn.click();
 
         //Assert
-        //TODO: Check other forms of text matchers
         home.alert.should('have.text', 
                     'This is just a dummy form, you just clicked SUBMIT BUTTON');
     });
 
+    //TODO: Make the it block text a data file
     it.skip('Go to EXPLORE LIFETIME MEMBERSHIP', () => {
         //Act
         home.exploreLifetimeMembershipLink.click();
@@ -107,10 +72,7 @@ describe('Demo Course - Search for a Course', () => {
     });
 
     //TODO: Make the it block text a data file
-    it('Slide Carousel to view Automation Architect Selenium with 7 live projects', () => {
-
-        cy.visit("https://www.way2automation.com/lifetime-membership-club/");
-
+    it.skip('Slide Carousel to view Automation Architect Selenium with 7 live projects', () => {
         //Arrange: Scroll to carousel section
         home.carouselSection.should('be.visible')
                             .scrollIntoView({ easing: 'linear', duration: 1000 }) //TODO: Do we need duration?
@@ -128,16 +90,21 @@ describe('Demo Course - Search for a Course', () => {
         cy.url().should('eq', 'https://www.selenium-tutorial.com/p/automation-architect-in-selenium-7-live-projects');
     });
 
-    it('Click Get Started', () => {
-        
-    });
+    it.skip('Search for the course CucumberParallelWithPageObjects - Project Code', () => {
+        const course = new CoursePage();
 
-    it('Search for the course CucumberParallelWithPageObjects - Project Code', () => {
-        
-    });
+        //Arrange
+        course.expandButton.click();
 
-    it('Should start', () => {
-        
+        //Act
+        cy.contains('CucumberParallelWithPageObjects - Project Code')
+                .should('be.visible')    
+                .scrollIntoView()
+                .contains('Start')
+                .click();
+
+        //Assert
+        course.heading.should('include.text', 'CucumberParallelWithPageObjects - Project Code')
     });
