+++ cypress/page-objects/course.page.ts
@@ -1,0 +1,15 @@
+
+
+export class CoursePage{
+
+
+    get expandButton(){
+        return cy.get('#more_lecture_sections');
+    }
+
+
+    get heading(){
+        return cy.get('#lecture_heading');
+    }
+
+}
